requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-csv-exports==1.0.5
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
magic-filter==1.0.9
Markdown==3.4.1
multidict==6.0.4
numpy==1.24.1
pandas==1.5.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
